// src/modules/iE/reports/thread-consumption-report/thread-consumption-report-type.ts
export type ThreadConsumptionReportType = {
  THREADCONSUMPTIONNO: string;
  STYLEID: number;
  COLORID: number;
  SIZEID: number;
  THICKNESS: number;
  THICKESSUOMID: number;
  WASTAGEPERCENT: number;
  TOTALOPERATIONLENGTH: number;
  TOTALWASTAGELENGTH: number;
  ACTUALOPERATIONLENGTH: number;
  CREATEBY: string;
  CREATEDATE: string;
  UPDATEBY: string;
  UPDATEDATE: string;
  STYLENO: string;
  STYLENAME: string;
  BUYERNAME: string;
  SIZENAME: string;
  COLORNAME: string;
  OPERATIONID: number;
  MACHINEID: number;
  STITCHID: number;
  STITCHNAME: string;
  ROWSOFSTITCH: number;
  SPINO: string;
  SEAMLENGTH: number;
  NEDDLETHREADID: number;
  NEDDLETHREADCOLORDEPENDENT: number;
  NEEDLETHREADCOLOR: string;
  BOBBINTHREADID: number;
  BOBBINTHREADCOLORDEPENDENT: number;
  BOBBINTHREADCOLOR: string;
  LOOPERTHREADID: number;
  LOOPERTHREADCOLORDEPENDENT: number;
  LOOPERTHREADCOLOR: string;
  NEEDLETHREADLENGTH: number;
  BOBBINTHREADLENGTH: number;
  LOOPERTHREADLENGTH: number;
  TOTALDETAILSOPERAITONLENGTH: number;
  WASTAGE: number;
  WASTAGEVALUE: number;
  OPERATIONNAME: string;
  MACHINECODE: string;
  MACHINENAME: string;
  CONE: string;
  NEDDLETHREADCODE: string;
  NEDDLETHREADNAME: string;
  LOOPERTHREADCODE: string;
  LOOPERTHREADNAME: string;
  BOBBINTHREADCODE: string;
  BOBBINTHREADNAME: string;
  COMPANYNAME: string;
  COMPANYADDRESS: string;
  COMPANYUTILITITY: string;
  ITEM_TYPE: string;
  SAMPLE_TYPE: string;
  GSM: string;
  MAIN_FABRIC: string;
  SIZE_LIST: string;
};